% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aqua.R
\name{theme_aqua}
\alias{theme_aqua}
\title{A custom theme for ggplot2 following the aqua report style for figures}
\usage{
theme_aqua(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base family}
}
\description{
A custom theme for ggplot2 following the aqua report style for figures
}
